Moved matchpy_connector to external
